requirements.txt
amqp==2.5.2
anyjson==0.3.3
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-celery==3.3.1
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-redis==4.11.0
django-rest-framework==0.1.0
django-reversion==3.0.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
flower==0.9.4
future==0.18.2
httplib2==0.9.2
humanize==0.5.1
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
kombu==4.6.8
librouteros==3.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
paramiko==2.7.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
supervisor==4.1.0
tablib==1.1.0
tornado==6.0.4
traitlets==4.3.3
vine==1.3.0
wcwidth==0.1.9
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
